Produces a surface based on an alternative ring-search strategy

Produces a surface based on an alternative ring-search strategy (ie searching in an expanding radius out from the 'crimes'). The output from this function can be used with geoProfile() and geoReportHitscores() to produce a map and hitscores based on this strategy.

Usage

geoRing(params, data, source, mcmc)

Arguments

params

Parameters list in the format defined by geoParams().

data

Data object in the format defined by geoData().

source

Potential sources object in the format defined by geoDataSource().

mcmc

mcmc object of the form produced by geoMCMC().

Value

A matrix containing a surface based on a ring-search strategy.

Examples

if (FALSE) { # interactive()
# \donttest{
# John Snow cholera data
d <- geoData(Cholera$longitude, Cholera$latitude)
s <- geoDataSource(WaterPumps$longitude, WaterPumps$latitude)
p <- geoParams(data = d, sigma_mean = 1.0, sigma_squared_shape = 2)
m <- geoMCMC(data = d, params = p)
surface_ring <- geoRing(params = p, data = d, source = s, mcmc = m)
gp_ring <- geoProfile(surface = surface_ring)
map <- geoPlotLeaflet(params = p,
                      data = d,
                      source = s,
                      surface = gp_ring,
                      opacity = 1)
map
# }
}